requirements.txt
streamlit==1.31.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.2.0
protobuf==3.20.3
